move funcitons around
