+++ models/demos/deepseek_v3_b1/fused_ops/attention_block/op.py
@@ -12,6 +12,10 @@ from models.demos.deepseek_v3_b1.blitz_decode_weights import (
     KVB12_PROJ_SingleDeviceOverlapSpec,
     O_PROJ_GATE_MM_RMSNORM_GAMMA_SingleDeviceOverlapSpec,
 )
+from models.demos.deepseek_v3_b1.circular_buffer_utils import (
+    cb_descriptor_from_overlapped_tensor,
+    cb_descriptor_from_overlapped_tensors,
+)
 from models.demos.deepseek_v3_b1.fused_ops.post_sdpa.op import _extend_runtime_args, _get_element_size_bytes, _round_up
 from models.demos.deepseek_v3_b1.micro_ops.flash_mla.op import (
     FlashMLADecode,
@@ -24,11 +28,7 @@ from models.demos.deepseek_v3_b1.unified_kernel_descriptor import (
     UnifiedCompileTimeCoreDescriptor,
     UnifiedKernelDescriptor,
 )
-from models.demos.deepseek_v3_b1.utils import (
-    cb_descriptor_from_overlapped_tensor,
-    cb_descriptor_from_overlapped_tensors,
-    float_to_uint32,
-)
+from models.demos.deepseek_v3_b1.utils import float_to_uint32
 
 
 def extend_fabric_args(existing_rt_args, fabric_args):
